overhaul Option, fixes incorrect new value loading

+++ src/main/java/net/geforcemods/securitycraft/renderers/TileEntitySecurityCameraRenderer.java
@@ -61,7 +61,7 @@ public class TileEntitySecurityCameraRenderer extends TileEntitySpecialRenderer<
 
 		GlStateManager.rotate(180F, rotation, 0.0F, 1.0F);
 
-		modelSecurityCamera.cameraRotationPoint.rotateAngleY = te.cameraRotation;
+		modelSecurityCamera.cameraRotationPoint.rotateAngleY = (float)te.cameraRotation;
 
 		modelSecurityCamera.render(null, 0.0F, 0.0F, -0.1F, 0.0F, 0.0F, 0.0625F);
